+++ R/MsBackendMsp.R
@@ -30,7 +30,9 @@ NULL
 #'
 #' @param format For `spectraVariableMapping`: `character(1)` specifying for
 #'     which MSP *flavour* the mapping should be returned. Currently supported
-#'     are: `format = "msp"` (generic MSP format).
+#'     are: `format = "msp"` (generic MSP format, for example for MS-DIAL MSP
+#'     files), `format = "lipidblast"` (MSP files in LipidBlast flavour) and
+#'     `format = "mona"` (MSP files in MoNA flavour).
 #' 
 #' @param mapping named `character` vector to rename MSP fields to spectra
 #'     variables (see [spectraVariableMapping()]). This allows to correctly
@@ -66,8 +68,29 @@ NULL
 #' be$intensity
 #' be$mz
 #'
+#' ## precursor m/z are however all missing
+#' be$precursorMz
+#'
 #' ## Default spectra variable mapping
 #' spectraVariableMapping(MsBackendMsp())
+#'
+#' ## In fact, to read MSP files in "LipidBlast flavour" we should use a
+#' ## different spectra variable mapping
+#' spectraVariableMapping(MsBackendMsp(), "lipidblast")
+#'
+#' ## Importing the data with this will correctly retrieve data
+#' be <- backendInitialize(MsBackendMsp(), f,
+#'     mapping = spectraVariableMapping(MsBackendMsp(), "lipidblast"))
+#' be$precursorMz
+#'
+#' ## Other fields are also correctly mapped, but might need to be converted
+#' ## to e.g. numeric, such as "exactmass"
+#' be$exactmass
+#'
+#' be$exactmass <- as.numeric(be$exactmass)
+#'
+#' be$adduct
+#' be$formula
 NULL
 
 setClass("MsBackendMsp",
@@ -131,7 +154,7 @@ MsBackendMsp <- function() {
 #'
 #' @rdname MsBackendMsp
 setMethod("spectraVariableMapping", "MsBackendMsp",
-          function(object, format = c("msp")) {
+          function(object, format = c("msp", "lipidblast", "mona")) {
               switch(match.arg(format),
                      "msp" = c(
                          name = "NAME",
@@ -142,6 +165,27 @@ setMethod("spectraVariableMapping", "MsBackendMsp",
                          exactmass = "EXACTMASS",
                          rtime = "RETENTIONTIME",
                          precursorMz = "PRECURSORMZ"
+                     ),
+                     "lipidblast" = c(
+                         name = "Name",
+                         synonym = "Synon",
+                         accession = "DB#",
+                         inchikey = "InChIKey",
+                         adduct = "Precursor_type",
+                         precursorMz = "PrecursorMZ",
+                         polarity = "Ion_mode",
+                         formula = "Formula",
+                         exactmass = "ExactMass"
+                     ),
+                     "mona" = c(
+                         name = "Name",
+                         synonym = "Synon",
+                         accession = "DB#",
+                         inchikey = "InChIKey",
+                         adduct = "Precursor_type",
+                         precursorMz = "PrecursorMZ",
+                         formula = "Formula",
+                         exactmass = "ExactMass"
                      )
                      )
           })